Separated domain calculation from zoom and general zoom refactoring
