--- src/assets/js/behavior/controlZoom.js
@@ -1,4 +1,0 @@
-export default function(zoomExtent) {
-    // If zooming, and about to pan off screen, do nothing
-    return (zoomExtent[0] > 0 && zoomExtent[1] < 0);
-}
