+++ src/assets/js/behavior/zoom/calculateDomain.js
@@ -1,0 +1,15 @@
+import util from '../../util/util';
+import zoomUtils from './zoomUtils';
+
+export default function(scale, data, discontinuityProvider, xExtent, zoomed, trackingLatest) {
+    var domain = scale.domain();
+
+    if (zoomed && trackingLatest) {
+        domain = util.domain.moveToLatest(
+            discontinuityProvider,
+            domain,
+            data);
+    }
+
+    return zoomUtils.clampDomain(scale, domain, data, xExtent);
+}
