+++ src/assets/js/behavior/zoom/clampDomain.js
